[131661] Define struct separately
